Fixed up extending SFX count on sounds.ini, now we can have more than 91 sounds!

+++ LunaDll/Globals.cpp
@@ -129,4 +129,4 @@ void printBoxA(const char *fmt, ...)
     dbgboxA(buf);
 }
 
-int soundEffectCount = 0;
+double soundEffectCount = 0;
